refactor(#3038): move tmux watcher turn identity decisions
